+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## dev
+
+- Sort table by the packages that have an update available
+
 ## [0.3.1]
 
 ### Added